+++ Assets/01Scenes/GameScene/Chapter1/Chapter1-Stage3.unity
@@ -792,15 +792,15 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 2136977446462511486, guid: 42f2af4b86912a34f9a911cc3828a686, type: 3}
       propertyPath: m_LocalPosition.x
-      value: -9.793724
+      value: -25.033726
       objectReference: {fileID: 0}
     - target: {fileID: 2136977446462511486, guid: 42f2af4b86912a34f9a911cc3828a686, type: 3}
       propertyPath: m_LocalPosition.y
-      value: 5.8000007
+      value: 5.8300014
       objectReference: {fileID: 0}
     - target: {fileID: 2136977446462511486, guid: 42f2af4b86912a34f9a911cc3828a686, type: 3}
       propertyPath: m_LocalPosition.z
-      value: -12.633724
+      value: -18.123724
       objectReference: {fileID: 0}
     - target: {fileID: 2136977446462511486, guid: 42f2af4b86912a34f9a911cc3828a686, type: 3}
       propertyPath: m_LocalRotation.w
@@ -1175,18 +1175,21 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 27955667c8f44f64883e75324ca27757, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  cutSceneType: 12
   chapterManager: {fileID: 0}
   cutScene: {fileID: 177563867}
+  cutSceneType: 12
+  isUseScripting: 1
+  scriptingList:
+  - {fileID: 11400000, guid: d227d11d98f0da447b2c19f5edcd2160, type: 2}
+  - {fileID: 11400000, guid: 4540b9a9bcdc32244b1bf28516d8a69a, type: 2}
+  isUseSkeleton: 0
+  skeletonParent: {fileID: 0}
   spawnerManager:
   - {fileID: 1391174613}
   - {fileID: 1721958487}
   interactionCol:
   - {fileID: 925382346}
   - {fileID: 1183940228}
-  scriptsList:
-  - {fileID: 11400000, guid: d227d11d98f0da447b2c19f5edcd2160, type: 2}
-  - {fileID: 11400000, guid: 4540b9a9bcdc32244b1bf28516d8a69a, type: 2}
 --- !u!1 &180463966
 GameObject:
   m_ObjectHideFlags: 0
@@ -2569,8 +2572,13 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 5b88b9dc1add20f44b1d7977eeac55c5, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  cutSceneType: 1
   chapterManager: {fileID: 0}
+  cutScene: {fileID: 0}
+  cutSceneType: 1
+  isUseScripting: 0
+  scriptingList: []
+  isUseSkeleton: 0
+  skeletonParent: {fileID: 0}
 --- !u!320 &632934173
 PlayableDirector:
   m_ObjectHideFlags: 0
@@ -4926,8 +4934,13 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: cad82560bb25010468300e9d2c3f42e4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  cutSceneType: 2
   chapterManager: {fileID: 0}
+  cutScene: {fileID: 0}
+  cutSceneType: 2
+  isUseScripting: 0
+  scriptingList: []
+  isUseSkeleton: 0
+  skeletonParent: {fileID: 0}
   grayScaleTime: 0.5
 --- !u!320 &1137756667
 PlayableDirector:
@@ -5881,13 +5894,16 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 2f169a833aad71f409f6ebbf06cd8101, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  cutSceneType: 7
   chapterManager: {fileID: 0}
-  chapter1Director: {fileID: 1431025555}
-  spawnerManager: {fileID: 1914705763}
-  scriptsList:
+  cutScene: {fileID: 0}
+  cutSceneType: 7
+  isUseScripting: 1
+  scriptingList:
   - {fileID: 11400000, guid: 3daef1e6e9a117f43b8de5a2d1b43018, type: 2}
   - {fileID: 11400000, guid: 3d7aa724d1777a446b134f47a30e6a42, type: 2}
+  isUseSkeleton: 0
+  skeletonParent: {fileID: 0}
+  spawnerManager: {fileID: 1914705763}
 --- !u!1 &1439828583
 GameObject:
   m_ObjectHideFlags: 0
@@ -8982,7 +8998,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 3953504714333392600, guid: 58b6e6a5adfbd3d499c6bc920614d15a, type: 3}
       propertyPath: m_IsActive
-      value: 1
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 3953504714333392601, guid: 58b6e6a5adfbd3d499c6bc920614d15a, type: 3}
       propertyPath: m_RootOrder
@@ -11516,8 +11532,13 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 518d2b5b2c533674697cda5ddffd84a7, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  cutSceneType: 3
   chapterManager: {fileID: 0}
+  cutScene: {fileID: 0}
+  cutSceneType: 3
+  isUseScripting: 0
+  scriptingList: []
+  isUseSkeleton: 0
+  skeletonParent: {fileID: 0}
 --- !u!320 &2086130054
 PlayableDirector:
   m_ObjectHideFlags: 0
@@ -11643,7 +11664,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 1355977930}
-  m_RootOrder: 4
+  m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!65 &2112876575
 BoxCollider:
